==> Rmd/03_Monocle3.Rmd <==
---
title: "Monocle3"
author: "Friederike Dündar"
date: "8/12/2019"
output:
  pdf_document:
    toc: yes
  html_document:
    toc: yes
    code_folding: hide
editor_options:
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, cache = TRUE)
```

```
conda create --name monocle --clone biocdevel
conda install -c r r-units 
conda install -c r r-lmtest  
conda install -c r r-sf
conda install -c r r-classInt
conda install -c r r-spdep
conda install -c r r-matrix 

~/miniconda3/envs/monocle/bin/R

install.packages("reticulate")
reticulate::py_install("louvain")

#devtools::install_github('cole-trapnell-lab/monocle3') ## couldn't get this to work

# cloned the repo
# in R: 
devtools::load_all("monocle3")

```

## 3dpa, all cells

```{r monocle_prep_3dpa_allClusters, eval=FALSE}
#screen -r jingli
# conda activate monocle
# ~/miniconda3/envs/monocle/bin/R
# cd 2019-08_Monocle/
devtools::load_all("monocle3/")
```

## 3dpa, all cells

```{r monocle_prep_3dpa_allClusters, eval=FALSE}
#screen -r jingli
# conda acticate monocle
# ~/miniconda3/envs/monocle/bin/R
# cd 2019-08_Monocle
devtools::load_all("monocle3/")
load("data/sce_CellGeneFilt_2019-06.rda")

## First, let's use all 3dpa cells (including those of non-Epi clusters)
cts <- counts(sce.filt[, sce.filt$condition == "3dpa"])
cd <- colData(sce.filt[, sce.filt$condition == "3dpa"])[, c("Sample", "ori_barcodes", "cell_names", "condition")]  

## FOLLOWING THESE PRESCRIPTIONS: https://cole-trapnell-lab.github.io/monocle3/monocle3_docs/#getting-started-with-monocle-3
cds <- new_cell_data_set(cts, cell_metadata = cd, gene_metadata = rowData(sce.filt))

cds <- preprocess_cds(cds, num_dim = 100)

set.seed(123)
cds <- reduce_dimension(cds, reduction_method = "UMAP", preprocess_method = "PCA")
cds <- cluster_cells(cds, reduction_method = "UMAP") # can be accessed:  clusters(cds)
## dimensionality reduction
cds <- learn_graph(cds)
## order_cells() assigns cells a pseudotime value based on their projection on the
## principal graph learned in the ‘learn_graph’ function and the
## position of chosen root states. 
#cds <- order_cells(cds) ## this needs shiny unless we specify the root nodes of the trajectory graph

## extracting root cells in a somewhat automated manner
get_earliest_principal_node <- function(cds, partition=NULL){
    
    if(!is.null(partition)){
        cell_ids <- which(partitions(cds) == partition)
    }else{
        cell_ids <- colnames(cds)
    }

  closest_vertex <-
    cds@principal_graph_aux[["UMAP"]]$pr_graph_cell_proj_closest_vertex
  closest_vertex <- as.matrix(closest_vertex[colnames(cds), ])
  
  root_pr_nodes <-
    igraph::V(principal_graph(cds)[["UMAP"]])$name[as.numeric(names
      (which.max(table(closest_vertex[cell_ids,]))))]

  root_pr_nodes
}

root_cells <- unlist(lapply(levels(partitions(cds)), function(x) get_earliest_principal_node(cds, partition = x)))
cds <- order_cells(cds, root_pr_nodes = root_cells)
saveRDS(cds, file = "monocle_3dpa_allCells.rds")
```

```{r plotting_cells_3dpa_all, eval=FALSE}
rowData(cds)$gene_short_name <- rowData(cds)$symbol

png("monocle3_3dpa_allCells.png", width = 700)
plot_cells(cds, color_cells_by = "cluster", ## this will be the clusters determined by monocle
           label_cell_groups=FALSE, label_leaves=FALSE,label_branch_points=FALSE,graph_label_size=5, 
           cell_size = 1)
dev.off()

png("monocle3_3dpa_allCells_markerGenes.png", width = 900, height = 900)
plot_cells(cds, genes = c("top2a","tcf21","fn1a", "cd74a","fli1a","tpm1"), label_cell_groups=FALSE, label_leaves=FALSE,label_branch_points=FALSE,graph_label_size=5, 
           cell_size = 1)
dev.off()
```

# Session info

```
R version 3.6.1 (2019-07-05)
Platform: x86_64-conda_cos6-linux-gnu (64-bit)
Running under: Red Hat Enterprise Linux Server release 6.3 (Santiago)
  
Matrix products: default
BLAS/LAPACK: /miniconda3/envs/monocle/lib/libopenblasp-r0.3.6.so
  
locale:
 [1] LC_CTYPE=en_US.UTF-8       LC_NUMERIC=C
 [3] LC_TIME=en_US.UTF-8        LC_COLLATE=en_US.UTF-8
 [5] LC_MONETARY=en_US.UTF-8    LC_MESSAGES=en_US.UTF-8
 [7] LC_PAPER=en_US.UTF-8       LC_NAME=C
 [9] LC_ADDRESS=C               LC_TELEPHONE=C
[11] LC_MEASUREMENT=en_US.UTF-8 LC_IDENTIFICATION=C       

attached base packages:
[1] stats4    parallel  stats     graphics  grDevices utils     datasets 
[8] methods   base     

other attached packages:
 [1] monocle3_0.1.3              testthat_2.2.1             
 [3] SingleCellExperiment_1.7.4  SummarizedExperiment_1.15.6
 [5] DelayedArray_0.11.4         BiocParallel_1.19.0        
 [7] matrixStats_0.54.0          GenomicRanges_1.37.14      
 [9] GenomeInfoDb_1.21.1         IRanges_2.19.10            
[11] S4Vectors_0.23.17           Biobase_2.45.0             
[13] BiocGenerics_0.31.5        

loaded via a namespace (and not attached):
  [1] Rtsne_0.15               colorspace_1.4-1         deldir_0.1-23           
  [4] grr_0.9.5                pryr_0.1.4               class_7.3-15            
  [7] rprojroot_1.3-2          XVector_0.25.0           fs_1.3.1                
 [10] rstudioapi_0.10          proxy_0.4-23             listenv_0.7.0           
 [13] furrr_0.1.0              remotes_2.1.0            ggrepel_0.8.1           
 [16] splines_3.6.1            codetools_0.2-16         pscl_1.5.2              
 [19] pkgload_1.0.2            speedglm_0.3-2           jsonlite_1.6            
 [22] RhpcBLASctl_0.18-205     uwot_0.1.3               pheatmap_1.0.12         
 [25] shiny_1.3.2              compiler_3.6.1           httr_1.4.1              
 [28] backports_1.1.4          assertthat_0.2.1         Matrix_1.2-17           
 [31] lazyeval_0.2.2           limma_3.41.15            cli_1.1.0               
 [34] later_0.8.0              htmltools_0.3.6          prettyunits_1.0.2       
 [37] tools_3.6.1              igraph_1.2.4.1           coda_0.19-3             
 [40] gtable_0.3.0             glue_1.3.1               GenomeInfoDbData_1.2.1  
 [43] RANN_2.6.1               reshape2_1.4.3           dplyr_0.8.3             
 [46] gmodels_2.18.1           Rcpp_1.0.2               slam_0.1-45             
 [49] spdep_1.1-2              gdata_2.18.0             nlme_3.1-141            
 [52] DelayedMatrixStats_1.7.1 lmtest_0.9-37            stringr_1.4.0           
 [55] globals_0.12.4           ps_1.3.0                 mime_0.7                
 [58] irlba_2.3.3              gtools_3.8.1             devtools_2.1.0          
 [61] future_1.14.0            LearnBayes_2.15.1        zlibbioc_1.31.0         
 [64] MASS_7.3-51.4            zoo_1.8-6                scales_1.0.0            
 [67] promises_1.0.1           expm_0.999-4             RColorBrewer_1.1-2      
 [70] gridExtra_2.3            memoise_1.1.0            reticulate_1.13         
 [73] pbapply_1.4-1            ggplot2_3.2.1            Matrix.utils_0.9.7      
 [76] stringi_1.4.3            desc_1.2.0               e1071_1.7-2             
 [79] boot_1.3-23              pkgbuild_1.0.4           spData_0.3.0            
 [82] rlang_0.4.0              pkgconfig_2.0.2          bitops_1.0-6            
 [85] rsample_0.0.5            lattice_0.20-38          purrr_0.3.2             
 [88] sf_0.7-6                 htmlwidgets_1.3          processx_3.4.1          
 [91] tidyselect_0.2.5         plyr_1.8.4               magrittr_1.5            
 [94] R6_2.4.0                 generics_0.0.2           DBI_1.0.0               
 [97] pillar_1.4.2             withr_2.1.2              units_0.6-3             
[100] sp_1.3-1                 RCurl_1.95-4.12          tibble_2.1.3            
[103] crayon_1.3.4             plotly_4.9.0             viridis_0.5.1           
[106] usethis_1.5.1            grid_3.6.1               data.table_1.12.2       
[109] callr_3.3.1              digest_0.6.20            classInt_0.3-1          
[112] pbmcapply_1.5.0          xtable_1.8-4             tidyr_0.8.3             
[115] httpuv_1.5.1             RcppParallel_4.4.3       munsell_0.5.0           
[118] viridisLite_0.3.0        sessioninfo_1.1.1 
```
