==> Rmd/04_figures.Rmd <==
---
title: "Figures: DotPlots, UMAPs, Monocle"
author: "Friederike Dündar"
date: "08/08/2021; updated `r Sys.Date()`"
output:
  html_document:
    toc: yes
    code_folding: hide
  pdf_document:
    toc: yes
    toc_depth: 5
editor_options:
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, cache=TRUE, cache.lazy = FALSE)
```
```{r include=FALSE, cache=FALSE, message=FALSE}
#library(scater)
library(SingleCellExperiment)
library(data.table)
library(magrittr)
library(ggplot2) ; theme_set(theme_bw(base_size = 16))
library(patchwork)
library(scABC2)
suppressMessages(library(clusterProfiler))
```


# Marker gene plot with *all* cells

Not just Tcf21-hi

```{r}
## all cells
sce <- readRDS("sce_CellGeneFiltWithScTransformBatchCorrect_2019-06.rds")
sce$clusters <- factor(sce$res.03, levels = c(0:11), ordered = TRUE)

cluster_cols <- ABCutilities:::fx.get_palette_ABC("paired_pal")[1:length(unique(sce$res.03))]
names(cluster_cols) <- levels(sce$res.03)

condition_cols <- c("cyan4","blue2","gray85")
names(condition_cols) <- as.character(unique(sce$condition))
```

```{r rev_allClusters_dot_plot_selected_markers, fig.height = 9.5, message=FALSE, warning = FALSE}
gns2 <-  c("tcf21", "gata5", "wt1b", "tbx18", "sema3d", "aldh1a2", "tgm2b", "fn1a",
    "serpinh1a", "serpinh1b", "col11a1a", "col12a1b", "ptx3a", "marcksb",
    ## additional genes:
    "atp5mc1", "hmgb2a", "hmgb2b", "psmb1","hsp90b1","calr3b","loxa",
    "hapln1a", "vegfaa", "clu", "cxcl12a", "col18a1a", "top2a", "pcna",
    "pdgfrb", "myh11a", "notch3", "tpm1", "acta2", "myl7", "tnnt2a",
    "mb", "kdrl", "fli1a", "mpeg1.1", "cd74a")
mks_mn2 <- scABC2::prep_dot_plot(sce,
    genes = gns2, exprs_values = "log1p_sctransform.batch",
    group_by = "res.03")
setorder(mks_mn2, res.03)

ggplot(mks_mn2, aes(x = factor(gene, levels = gns2, ordered = TRUE),
    y = res.03, color = mean.norm.exprs, size = pct.expressed)) +
    geom_point() + coord_flip() +
    scale_color_gradient(low = "grey95", high = "red3", 
      limits = c(0, max(mks_mn2$mean.norm.exprs))) +
    ylab("cluster") + 
    #ggtitle("marker genes of epicardial, CM/SMC, endothelial and blood cells")+
    xlab("")
```

# UMAPs

## Fig a

```{r rev_umap_allCells_condition, fig.height = 7, fig.width = 9.5}
scABC2::plot_reducedDim_from_sce(sce,
    which_reddim = "umap_scTransformBatch", which_pcs = c(1:2),
    exprs_values = "log1p_sctransform.batch",
    color_by = "condition", 
    label_by = "res.03", label_size = 4, alpha = .3,
    set_colors = TRUE,
    size_by = .5, remove_rug = TRUE) 
```


## Fig c

```{r rev_umap_allCells_clusters, fig.height = 7, fig.width = 8}
scABC2::plot_reducedDim_from_sce(sce.object = sce,
    which_reddim = "umap_scTransformBatch", which_pcs = c(1:2),
    exprs_values = "log1p_sctransform.batch",
    color_by = "res.03", 
    add_cell_info = "condition",label_by = "res.03",
    label_size = 4,
    alpha = .3, set_colors = TRUE,
    size_by = .5, remove_rug = TRUE) +
    theme(legend.position = "bottom") 
```

## Fig e

```{r rev_umap_allCells_clusters_per_condition, fig.width = 21, fig.height = 9}
scABC2::plot_reducedDim_from_sce(sce,
    which_reddim = "umap_scTransformBatch", which_pcs = c(1:2),
    exprs_values = "log1p_sctransform.batch",
    color_by = "res.03", 
    add_cell_info = "condition",
    label_by = "res.03", label_size = 4,
    alpha = .3, set_colors = TRUE,
    size_by = .5, remove_rug = TRUE) +
    theme(legend.position = "bottom") +
    facet_wrap(~condition)
```


## Monocle3 with pseudotime

```{r eval=FALSE}
#screen -r jingli
# conda activate monocle
# ~/miniconda3/envs/monocle/bin/R
devtools::load_all("monocle3/")

cds <- readRDS(file = "monocle_3dpa_allCells.rds")

pdf("monocle3_3dpa_allCells_pseudotime.pdf", width = 15, height = 15)
plot_cells(cds,
           color_cells_by = "pseudotime",
           label_cell_groups=FALSE,
           label_leaves=FALSE,
           label_branch_points=FALSE,
           graph_label_size=5, cell_size=1)
dev.off()
```
